File Organizations and Codebase Cleanup
