+++ Gista/Shared/Models/Gists/Gist.swift
@@ -150,37 +150,6 @@ public struct Gist: Identifiable, Codable {
     }
 }
 
-// MARK: - Preview Data
-public extension Gist {
-    static let previews: [Gist] = [
-        Gist(
-            title: "Tech Articles",
-            category: "Technology",
-            imageUrl: "https://example.com/image.jpg",
-            link: "https://example.com/tech",
-            playbackDuration: 120,
-            publisher: "TechGista",
-            segments: [
-                GistSegment(duration: 60, title: "Intro", audioUrl: "https://example.com/audio1.mp3")
-            ],
-            status: GistStatus(inProduction: false, productionStatus: "Reviewing Content"),
-            color: .blue
-        ),
-        Gist(
-            title: "Business",
-            category: "Business",
-            imageUrl: "https://example.com/image.jpg",
-            link: "https://example.com/business",
-            playbackDuration: 180,
-            publisher: "BizGista",
-            segments: [
-                GistSegment(duration: 90, title: "Market Update", audioUrl: "https://example.com/audio2.mp3")
-            ],
-            status: GistStatus(inProduction: false, productionStatus: "Reviewing Content"),
-            color: .green
-        )
-    ]
-}
 
 // MARK: - Core Supporting Types
 public struct GistSegment: Codable {
@@ -277,138 +246,6 @@ public struct GistStatus: Codable {
     }
 }
 
-public struct GistCreatedStatus: Codable {
-    public let gistCreated: Bool
-    public let gistId: String?
-    public let imageUrl: String?
-    public let linkId: String
-    public let linkTitle: String
-    public let linkType: String
-    public let url: String
-    
-    public init(
-        gistCreated: Bool,
-        gistId: String?,
-        imageUrl: String?,
-        linkId: String,
-        linkTitle: String,
-        linkType: String,
-        url: String
-    ) {
-        self.gistCreated = gistCreated
-        self.gistId = gistId
-        self.imageUrl = imageUrl
-        self.linkId = linkId
-        self.linkTitle = linkTitle
-        self.linkType = linkType
-        self.url = url
-    }
-    
-    enum CodingKeys: String, CodingKey {
-        case gistCreated = "gist_created"
-        case gistId = "gist_id"
-        case imageUrl = "image_url"
-        case linkId = "link_id"
-        case linkTitle = "link_title"
-        case linkType = "link_type"
-        case url
-    }
-}
-
-// MARK: - API Models
-public struct GistRequest: Codable {
-    public let title: String
-    public let link: String
-    public let imageUrl: String
-    public let category: String
-    public let segments: [GistSegment]
-    public let playbackDuration: Int
-    public let linkId: String
-    public let gistId: String?
-    public let isFinished: Bool
-    public let playbackTime: Int
-    public let status: GistStatus
-    
-    public init(
-        title: String,
-        link: String,
-        imageUrl: String,
-        category: String,
-        segments: [GistSegment],
-        playbackDuration: Int,
-        linkId: String,
-        gistId: String? = nil,
-        isFinished: Bool = false,
-        playbackTime: Int = 0,
-        status: GistStatus = GistStatus(inProduction: false, productionStatus: "Reviewing Content")
-    ) {
-        self.title = title
-        self.link = link
-        self.imageUrl = imageUrl
-        self.category = category
-        self.segments = segments
-        self.playbackDuration = playbackDuration
-        self.linkId = linkId
-        self.gistId = gistId
-        self.isFinished = isFinished
-        self.playbackTime = playbackTime
-        self.status = status
-    }
-    
-    enum CodingKeys: String, CodingKey {
-        case title, link, category, segments, status, isFinished
-        case imageUrl = "image_url"
-        case playbackDuration = "playback_duration"
-        case linkId = "link_id"
-        case gistId = "gist_id"
-        case playbackTime = "playback_time"
-    }
-    
-    // Custom encode method to ensure proper formatting for the API
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        // Ensure title is never undefined
-        let safeTitle = title.isEmpty ? "Untitled Gist" : title
-        try container.encode(safeTitle, forKey: .title)
-        
-        // Ensure link is never undefined
-        let safeLink = link.isEmpty ? "https://example.com" : link
-        try container.encode(safeLink, forKey: .link)
-        
-        // Ensure imageUrl is never undefined
-        let safeImageUrl = imageUrl.isEmpty ? "https://example.com/image.jpg" : imageUrl
-        try container.encode(safeImageUrl, forKey: .imageUrl)
-        
-        // Ensure category is never undefined
-        let safeCategory = category.isEmpty ? "General" : category
-        try container.encode(safeCategory, forKey: .category)
-        
-        // Ensure segments array is never empty
-        if segments.isEmpty {
-            let defaultSegment = GistSegment(
-                duration: 60,
-                title: "Default Segment",
-                audioUrl: "https://example.com/audio.mp3",
-                segmentIndex: 0
-            )
-            try container.encode([defaultSegment], forKey: .segments)
-        } else {
-            try container.encode(segments, forKey: .segments)
-        }
-        
-        try container.encode(playbackDuration, forKey: .playbackDuration)
-        try container.encode(linkId, forKey: .linkId)
-        
-        if let gistId = gistId {
-            try container.encode(gistId, forKey: .gistId)
-        }
-        
-        try container.encode(isFinished, forKey: .isFinished)
-        try container.encode(playbackTime, forKey: .playbackTime)
-        try container.encode(status, forKey: .status)
-    }
-}
 
 // MARK: - UI Supporting Types
 public struct GistColor: Codable {
@@ -442,34 +279,3 @@ public struct GistColor: Codable {
     ]
 }
 
-// MARK: - Color Extension
-extension Color: Codable {
-    enum CodingKeys: String, CodingKey {
-        case red, green, blue, alpha
-    }
-    
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        let r = try container.decode(Double.self, forKey: .red)
-        let g = try container.decode(Double.self, forKey: .green)
-        let b = try container.decode(Double.self, forKey: .blue)
-        let a = try container.decode(Double.self, forKey: .alpha)
-        
-        self.init(red: r, green: g, blue: b, opacity: a)
-    }
-    
-    public func encode(to encoder: Encoder) throws {
-        guard let colorComponents = self.cgColor?.components else {
-            throw EncodingError.invalidValue(self, EncodingError.Context(
-                codingPath: encoder.codingPath,
-                debugDescription: "Could not get color components"
-            ))
-        }
-        
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(colorComponents[0], forKey: .red)
-        try container.encode(colorComponents[1], forKey: .green)
-        try container.encode(colorComponents[2], forKey: .blue)
-        try container.encode(colorComponents[3], forKey: .alpha)
-    }
-}
